refactor(config-ui): adjust the type position for create bp
