+++ config-ui/src/pages/blueprint/create/bp-context.tsx
@@ -23,11 +23,10 @@ import type { ConnectionItemType } from '@/store'
 import { useConnection, ConnectionStatusEnum } from '@/store'
 import { operator } from '@/utils'
 
-import { ModeEnum } from '../types'
+import { ModeEnum, FromEnum } from '../types'
 import { validRawPlan } from '../utils'
 
 import type { BPContextType } from './types'
-import { FromEnum } from './types'
 import * as API from './api'
 
 export const BPContext = React.createContext<BPContextType>({
